Customize scrollbar styles for improved UI consistency
Updated the global styles to introduce custom styling for the web page's
scrollbar. The changes include specifying the width of the scrollbar,
customizing its track and thumb components with new colors and a rounded
appearance for better visual integration with the site's theme.
Additionally, we have a hover effect to enhance user interaction. These
custom scrollbar styles also enhance the overall user experience by
providing a more cohesive and app-like feel to the site.